Token-Schema corrigido
