+++ app/api/auth.py
@@ -6,40 +6,28 @@ from typing import Optional
 from fastapi import Depends, HTTPException, status
 from fastapi.security import OAuth2PasswordBearer
 from jose import JWTError, jwt
-from passlib.context import CryptContext
 from pydantic import BaseModel
-from app.models import user_model
-from app.schemas import user_schema
-from app.api.services import user_service
-from app.db.database import SessionLocal
+from sqlalchemy.orm import Session
 
-# --- Configuração de Senha ---
-# Usa o algoritmo bcrypt para hashing
-pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
+# Importa as funções de senha do novo módulo
+from app.api.core.security import verify_password
+
+# Importa os outros módulos necessários
+from app.db.database import SessionLocal
+from app.models import user_model
+from app.schemas import user_schema, token_schema
+from app.api.services import user_service
+
 
 # --- Configuração do JWT ---
-# Chave SECRETA para assinar o token. Em deploy guardar uma chave
-SECRET_KEY = "sua-chave-secreta-muito-forte-e-dificil-de-adivinhar"
+SECRET_KEY = "senha" # Senha para fazer o hash
 ALGORITHM = "HS256"
 ACCESS_TOKEN_EXPIRE_MINUTES = 30
 
-# OAuth2PasswordBearer é uma dependência que extrai o token do cabeçalho "Authorization"
 oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
 
-class TokenData(BaseModel):
-    username: Optional[str] = None
-
-
-def verify_password(plain_password: str, hashed_password: str) -> bool:
-    """Verifica se a senha fornecida corresponde ao hash."""
-    return pwd_context.verify(plain_password, hashed_password)
-
-def get_password_hash(password: str) -> str:
-    """Gera o hash de uma senha."""
-    return pwd_context.hash(password)
 
 def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
-    """Cria um novo token de acesso JWT."""
     to_encode = data.copy()
     if expires_delta:
         expire = datetime.now(timezone.utc) + expires_delta
@@ -60,10 +48,10 @@ def get_current_user(token: str = Depends(oauth2_scheme), db: Session = Depends(
         username: str = payload.get("sub")
         if username is None:
             raise credentials_exception
-        token_data = TokenData(username=username)
+        token_data = token_schema.TokenData(username=username)
     except JWTError:
         raise credentials_exception
-
+    
     user = user_service.get_user_by_username(db, username=token_data.username)
     if user is None:
         raise credentials_exception
