+++ app/api/core/security.py
@@ -1,0 +1,13 @@
+# app/api/core/security.py
+
+from passlib.context import CryptContext
+
+pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
+
+def verify_password(plain_password: str, hashed_password: str) -> bool:
+    """Verifica se a senha fornecida corresponde ao hash armazenado."""
+    return pwd_context.verify(plain_password, hashed_password)
+
+def get_password_hash(password: str) -> str:
+    """Gera o hash de uma senha."""
+    return pwd_context.hash(password)
