+++ requirements.txt
@@ -7,3 +7,4 @@ httpx
 passlib[bcrypt]
 python-jose[cryptography]
 email-validator
+python-multipart
